pyproject.toml
[project]
name = "gh-unneeded"
version = "1.0.0"
description = "Determine whether we don't need this clone (or some of its branches) anymore."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython>=3.1.46",
]

[project.scripts]
git-fork = "git_unneeded:main"

[build-system]
requires = ["setuptools >= 80.0", "setuptools-scm[simple]>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["git_unneeded"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mypy>=1.0.1",
]
readme = [
    "cogapp>=3.6.0",
]


[tool.pytest.ini_options]
addopts = "--last-failed --mypy --cov --no-cov-on-fail --failed-first"
strict_xfail = true

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.mypy]
check_untyped_defs = true
warn_unused_ignores = true
warn_unreachable = true
show_error_codes = true
warn_return_any = true
disallow_any_generics = true
disallow_untyped_defs = true

files = "git_unneeded.py,tests"


[tool.ruff]
line-length = 160

[tool.ruff.lint]
select = [
    "I",  # import sort
    "E",  # pycodestyle
    "F",  # pyflakes
    "UP",  # pyupgrade
]

[tool.ruff.lint.pycodestyle]
max-line-length = 160
max-doc-length = 300  # really long comments are OK

[tool.ruff.lint.per-file-ignores]
"test_*.py" = ["E501"]

# uv tool run pytest-watcher -sv .
[tool.pytest-watcher]
now = true
clear = true
notify_on_failure = false
delay = 0.2
runner = ".venv/bin/pytest"
patterns = ["*.py"]
ignore_patterns = []
